+++ src/plone/rest/tests/test_filestream_iterator.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from plone.app.testing import SITE_OWNER_NAME
+from plone.app.testing import SITE_OWNER_PASSWORD
+from plone.rest.testing import PLONE_REST_FUNCTIONAL_TESTING
+
+import unittest
+import requests
+import transaction
+
+
+class TestFilestreamIterator(unittest.TestCase):
+
+    layer = PLONE_REST_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer['app']
+        self.request = self.layer['request']
+        self.portal = self.layer['portal']
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.portal.invokeFactory('Document', id='doc1')
+        self.document = self.portal['doc1']
+        transaction.commit()
+
+    def test_filestream_iterator(self):
+        response = requests.get(
+            self.document.absolute_url() + '/filestreamiterator',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD))
+        # demo.py returning itself as content
+        self.assertTrue('-*- coding' in response.content)